Move handleBackgroundReplacement to AntMedia.js
